+++ backend/drivers/drivers/switch_linux.go
@@ -9,7 +9,6 @@ type LinuxSwitchDriver struct {
 	drivers_meta.Meta
 	capabilities.L2Base
 	capabilities.SwitchBase
-	capabilities.TCBase
 }
 
 func NewLinuxSwitchDriver() *LinuxSwitchDriver {
@@ -21,4 +20,3 @@ func NewLinuxSwitchDriver() *LinuxSwitchDriver {
 // Checks opcionales
 var _ capabilities.L2Capable = (*LinuxSwitchDriver)(nil)
 var _ capabilities.SwitchCapable = (*LinuxSwitchDriver)(nil)
-var _ capabilities.TCCapable = (*LinuxSwitchDriver)(nil)